change nofile limit value in SAP Note 1771258 to fit the HANA requirements. (even that we are still waiting for SAP to updtae the Note)
