+++ sap/note/sectLimits.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"github.com/SUSE/saptune/system"
 	"os"
+	"strconv"
 	"strings"
 )
 
@@ -11,7 +12,8 @@ import (
 
 // GetLimitsVal initialise the security limit structure with the current
 // system settings
-func GetLimitsVal(value string) (string, error) {
+func GetLimitsVal(value string) (string, string, error) {
+	info := ""
 	// Find out current limits
 	limit := value
 	if limit != "" && limit != "NA" {
@@ -21,11 +23,16 @@ func GetLimitsVal(value string) (string, error) {
 		// a valid limits entry
 
 		// /etc/security/limits.d/saptune-<domain>-<item>-<type>.conf
+		if len(lim) < 3 {
+			return "", info, fmt.Errorf("Wrong format of limit entry in Note definition file. Please check")
+		} else if len(lim) > 3 {
+			info = getLimitInfo(lim[3])
+		}
 		dropInFile := fmt.Sprintf("/etc/security/limits.d/saptune-%s-%s-%s.conf", lim[0], lim[2], lim[1])
 		secLimits, err := system.ParseSecLimitsFile(dropInFile)
 		if err != nil {
 			//ANGI TODO - check, if other files in /etc/security/limits.d contain a value for the touple "<domain>-<item>-<type>"
-			return "", err
+			return "", info, err
 		}
 		lim[3], _ = secLimits.Get(lim[0], lim[1], lim[2])
 		if lim[3] == "" {
@@ -34,7 +41,7 @@ func GetLimitsVal(value string) (string, error) {
 		// current limit found
 		limit = strings.Join(lim, " ")
 	}
-	return limit, nil
+	return limit, info, nil
 }
 
 // OptLimitsVal optimises the security limit structure with the settings
@@ -77,4 +84,17 @@ func SetLimitsVal(key, noteID, value string, revert bool) error {
 		}
 	}
 	return err
+}
+
+// getLimitInfo returns info, if config value exceeds system limit for nofile
+func getLimitInfo(val string) string {
+	info := ""
+	if val != "" && val != "NA" && val != "infinity" && val != "unlimited" && val != "-1" {
+		nrOpen, _ := system.GetSysctlInt("fs.nr_open")
+		ilim, _ := strconv.Atoi(val)
+		if ilim > nrOpen {
+			info = "limit_exceeded"
+		}
+	}
+	return info
 }
